+++ x/oracle/keeper/update_exchange_rates.go
@@ -20,7 +20,7 @@ func (k Keeper) UpdateExchangeRates(ctx sdk.Context) {
 
 	k.countVotesAndUpdateExchangeRates(ctx, pairBallotsMap, validatorPerformances)
 	k.registerMissedVotes(ctx, whitelistedPairs, validatorPerformances)
-	k.rewardBallotWinners(ctx, whitelistedPairs, validatorPerformances)
+	k.rewardBallotWinners(ctx, validatorPerformances)
 
 	params := k.GetParams(ctx)
 	k.clearVotesAndPreVotes(ctx, params.VotePeriod)